refactored the HTTPParser
It, as of now, properly handles everything related to parsing.
It should be fed with data via the insertion operator and the parsing starts
automatically, if there is an error doring parsing the
parseError(const QString &reason) signal is emitted or if the parse is succesful
parsed() is emitted. Please note that you can send data as it's received from
the network to the parser, since if it cannot parse what it already has, it will
wait for more data until it can parse it.